refactor: move resource funs into a new file
