+++ pkg/kvm/transport_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/opiproject/gospdk/spdk"
 	pb "github.com/opiproject/opi-api/storage/v1alpha1/gen/go"
-	"github.com/opiproject/opi-spdk-bridge/pkg/frontend"
+	"github.com/opiproject/opi-spdk-bridge/pkg/utils"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 )
 
@@ -108,7 +108,7 @@ func TestNewNvmeVfiouserTransportParams(t *testing.T) {
 		t.Run(testName, func(t *testing.T) {
 			vfiouserTransport := NewNvmeVfiouserTransport(tmpDir)
 			gotParams, err := vfiouserTransport.Params(&pb.NvmeController{
-				Name: frontend.ResourceIDToControllerName("subsys0", "nvme-1"),
+				Name: utils.ResourceIDToControllerName("subsys0", "nvme-1"),
 				Spec: &pb.NvmeControllerSpec{
 					Endpoint: &pb.NvmeControllerSpec_PcieId{
 						PcieId: &pb.PciEndpoint{
